🗃️ refactor : 디렉토리 구조 변경, NotFoundException에서 NOT_FOUND 에러코드 사용하도록 변경
